+++ tests/test_v092_legacy_path.py
@@ -78,11 +78,11 @@ class LegacyHost(MergedEvalMixin, DesireMixin, DangerMixin):
     async def _is_desire_already_expressed(self, d, r, event=None):
         return False
 
-    def _read_worldview(self):
+    def _read_worldview(self, umo=""):
         import copy
         return copy.deepcopy(self._worldview)
 
-    def _write_worldview(self, data):
+    def _write_worldview(self, data, umo=""):
         self._worldview = data
 
     def _read_desires(self):